Move types into own modules, for consistency
Also adds missing Haddock documentation.
