+++ package.yaml
@@ -317,10 +317,12 @@ library:
   - Stack.Types.GhcPkgId
   - Stack.Types.GlobalOpts
   - Stack.Types.GlobalOptsMonoid
+  - Stack.Types.HpcReportOpts
   - Stack.Types.Installed
   - Stack.Types.InterfaceOpt
   - Stack.Types.IsMutable
   - Stack.Types.LockFileBehavior
+  - Stack.Types.LsOpts
   - Stack.Types.MsysEnvironment
   - Stack.Types.NamedComponent
   - Stack.Types.Nix
