+++ stack.cabal
@@ -353,10 +353,12 @@ library
       Stack.Types.GhcPkgId
       Stack.Types.GlobalOpts
       Stack.Types.GlobalOptsMonoid
+      Stack.Types.HpcReportOpts
       Stack.Types.Installed
       Stack.Types.InterfaceOpt
       Stack.Types.IsMutable
       Stack.Types.LockFileBehavior
+      Stack.Types.LsOpts
       Stack.Types.MsysEnvironment
       Stack.Types.NamedComponent
       Stack.Types.Nix
@@ -395,6 +397,8 @@ library
       System.Terminal
       Build_stack
       Paths_stack
+  other-modules:
+      Stack.Types.SetupOpts
   autogen-modules:
       Build_stack
       Paths_stack
